This is a small optimization, which turns a seek + writev into a pwritev syscall on Linux.