@bmatsuo 2016-03-09
Oh interesting. I didn't realize that this system call was guarded by a `#define`.

Are there times when pwritev will actually hurt performance? It doesn't sound like it would. But I don't actually have much experience with the call. And right now I'm not sure why it would be placed behind a flag.

One of the few times I have seen the call discussed is in https://github.com/boltdb/bolt/issues/238. But that seems more like it wasn't implemented due to constrained resources than potential issues.

I will benchmark, look up anything I can find it the mailing list, and get back here.

Thanks