@bmatsuo 2016-03-12
I don't really see a change in performance. Do you actually see any change if you run the benchmarks or in any programs you have?

```
benchmark                              old ns/op     new ns/op     delta
BenchmarkEnv_ReaderList-4              51130         51319         +0.37%
BenchmarkTxn_Put-4                     1526          1530          +0.26%
BenchmarkTxn_PutReserve-4              1610          1611          +0.06%
BenchmarkTxn_PutReserve_writemap-4     1344          1335          -0.67%
BenchmarkTxn_Put_writemap-4            1278          1290          +0.94%
BenchmarkTxn_Get_ro-4                  1398          1416          +1.29%
BenchmarkTxn_Get_raw_ro-4              783           798           +1.92%
BenchmarkScan_ro-4                     5890150       5901033       +0.18%
BenchmarkScan_raw_ro-4                 1521376       1527676       +0.41%
BenchmarkCursor-4                      691           687           -0.58%
BenchmarkCursor_Renew-4                179           178           -0.56%
BenchmarkTxn_Sub_commit-4              189910        188408        -0.79%
BenchmarkTxn_Sub_abort-4               209512        184603        -11.89%
BenchmarkTxn_abort-4                   14391         14075         -2.20%
BenchmarkTxn_commit-4                  14413         14634         +1.53%
BenchmarkTxn_ro-4                      144813        161165        +11.29%
BenchmarkTxn_unmanaged_abort-4         14490         14374         -0.80%
BenchmarkTxn_unmanaged_commit-4        14510         14517         +0.05%
BenchmarkTxn_unmanaged_ro-4            149717        147595        -1.42%
BenchmarkTxn_renew-4                   532           528           -0.75%
BenchmarkTxn_Put_append-4              468           465           -0.64%
BenchmarkTxn_Put_append_noflag-4       619           653           +5.49%
```

The benchmarks above that really swing are unrelated, I think. They should not be flushing pages. And they are generally quite noisy.

I also haven't been able to find any discussion of the flag in the official mailing list. I may have missed something. But I'm really at a loss about this flag. Perhaps I don't have the appropriate use case in the benchmarks. In general I think the benchmarks could be improved and more focused. But they do provide some utility.