@mranney 2016-03-13
I'm surprised that this doesn't help. Seems like the cutting the number of system calls in half should be an obvious win. Maybe the writes in the bench loop are so fast as it is that the difference ends up not showing up compared to whatever else is going on.