@bmatsuo 2016-03-13
I have started writing some benchmarks which I believe will make the impact more apparent. And what I've seen so far makes me hopeful.

Looking over how pwritev gets used. It seems that you would see the flag's benefits when committing multiple writes (on multiple pages) per transaction. I don't think that this is being benchmarked effectively on master.