@bmatsuo 2016-03-13
@mranney furthermore, to your point about it being an obvious win, I would absolutely agree with you taking the system call and LMDB code at face value. The thing that confuses me is why LMDB will not try define this for linux systems by default.

The C Makefile states

> Note that the defaults [of MDB_USE_PWRITEV] should already be correct on most platforms; you should not need to change [its value].

AFAIK the library then make no attempt to define the flag for linux systems.