@lmb 2016-03-13
My unsubstantiated guess is that the flag is set in the OpenLDAP Makefile,
but I have not checked yet.
On 13 Mar 2016 11:54 am, "Bryan Matsuo" <notifications@github.com> wrote:

> @mranney <https://github.com/mranney> furthermore, to your point about it
> being an obvious win, I would absolutely agree with you taking the system
> call and LMDB code at face value. The thing that confuses me is why LMDB
> will not try define this for linux systems by default.
>
> The C Makefile states
>
> Note that the defaults [of MDB_USE_PWRITEV] should already be correct on
> most platforms; you should not need to change [its value].
>
> AFAIK the library then make no attempt to define the flag for linux
> systems.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/bmatsuo/lmdb-go/pull/58#issuecomment-195941833>.
>