@bmatsuo 2016-03-13
@lmb AFAICT that is not the case. It looks like there is a "grep" interface for the openldap repo.

http://www.openldap.org/devel/gitweb.cgi?p=openldap.git&a=search&h=HEAD&st=grep&s=PWRITEV

Searching for "PWRITEV" only turns up known references in the liblmdb tree.