@bmatsuo 2016-03-13
I've put the question to the openldap-technical mailing list. A definitive answer should come from there. I will link the response here from the web archives after I have received it.