@bmatsuo 2016-03-23
Sorry for not posting an update. But I never got a response from the openldap-technical mailing list. So I'm still a little confused there.

But it has come to my attention that the minimum system requirements for Go on Linux are kernel version 2.6.23. I was previously under the impression that the minimum required version was 2.6.32 (oops)

https://golang.org/doc/install#requirements

The pwrite(2) system call wasn't introduced until 2.6.30. So I'm somewhat concerned about people not being able to build the package if the have an old kernel.

I don't use custom build tags much. But would it be practical to put the `#CGO linux CFLAGS` in a separate file that is guarded by some special build tag like `pwritev`?

```
//+build pwritev

package lmdb

/*
CGO linux CFLAGS: -DMDB_USE_PWRITEV
*/
import C
``` 

That might not be completely correct syntax, but there is otherwise possible to define CFLAGS in multiple places AFAIK. I just don't really know the details of how to build non-main packages with custom build tags.

---

On another note I haven't made significant progress on benchmarks. I actually hate writing benchmarks for lmdb-go. While the current ones are especially bad it is impossible to completely remedy because they always be inherently artificial.